requirements.txt
streamlit
pandas
yfinance
pandas-ta-classic
newsapi-python
transformers
torch
google-genai
plotly
python-dotenv
requests